handling interpolation

+++ src/pyqentangle/continuous.py
@@ -7,8 +7,8 @@ import numba as nb
 from . import schmidt_decomposition, OutOfRangeException, UnequalLengthException
 
 
-@nb.njit(nb.float64(nb.float64[:], nb.float64[:], nb.float64))
-def interpolate(xarray: np.ndarray, yarray: np.ndarray, x: float) -> float:
+@nb.njit(nb.float64(nb.float64[:], nb.complex128[:], nb.float64))
+def interpolate(xarray: np.ndarray, yarray: np.ndarray, x: float) -> np.complex128:
     left = 0
     right = len(xarray) - 1
     idx = right // 2
